+++ test/ext-test/cpu/makie/makie_ext.jl
@@ -3,9 +3,10 @@
     xvec = yvec = -15.0:0.1:15.0
     wig = transpose(wigner(ψ, xvec, yvec))
 
+    # Makie unload errors
     @test_throws ArgumentError plot_wigner(ψ; library = :Makie, xvec = xvec, yvec = yvec)
-
     @test_throws ArgumentError plot_fock_distribution(ψ; library = :Makie)
+    @test_throws ArgumentError plot_bloch(ψ; library = :Makie)
 
     using Makie
 
@@ -63,12 +64,12 @@
 end
 
 @testset "Makie Bloch sphere" begin
-    ρ = 0.7*ket2dm(basis(2, 0)) + 0.3*ket2dm(basis(2, 1))
+    ρ = 0.7 * ket2dm(basis(2, 0)) + 0.3 * ket2dm(basis(2, 1))
     fig, ax = plot_bloch(ρ)
     @test fig isa Figure
     @test ax isa Axis3
 
-    ψ = (basis(2, 0) + basis(2, 1))/√2
+    ψ = (basis(2, 0) + basis(2, 1)) / √2
     fig, ax = plot_bloch(ψ)
     @test fig isa Figure
     @test ax isa Axis3
@@ -92,25 +93,37 @@ end
     add_points!(b, hcat(pts...))
     @test length(b.points) == 2
     @test b.points[2] ≈ hcat(pts...)
+    @test_throws ArgumentError add_points!(b, [1 2 3 4])
+    @test_throws ArgumentError add_points!(b, pts; meth = :wrong)
 
     b = Bloch()
     add_vectors!(b, [1.0, 1.0, 0.0])
     @test length(b.vectors) == 1
-    @test isapprox(norm(b.vectors[1]), 1.0)
+    @test isapprox(norm(b.vectors[1]), √2)
 
     vecs = [[0.0, 0.0, 1.0], [1.0, 0.0, 0.0]]
     add_vectors!(b, vecs)
     @test length(b.vectors) == 3
-    @test all(norm(v) ≈ 1.0 for v in b.vectors)
+    @test isapprox(norm(b.vectors[2]), 1.0)
+    @test isapprox(norm(b.vectors[3]), 1.0)
 
+    vec_correct = [1, 0, 0]
+    vec_wrong = [1, 0]
     b = Bloch()
     add_line!(b, [0, 0, 0], [1, 0, 0])
     @test length(b.lines) == 1
     @test b.lines[1][1][3] ≈ [0.0, 0.0]
+    @test_throws ArgumentError add_line!(b, vec_wrong, vec_correct)
+    @test_throws ArgumentError add_line!(b, vec_correct, vec_wrong)
 
     add_arc!(b, [0, 0, 1], [0, 1, 0], [1, 0, 0])
     @test length(b.arcs) == 1
     @test b.arcs[1][3] == [1.0, 0.0, 0.0]
+    @test_throws ArgumentError add_arc!(b, vec_wrong, vec_correct)
+    @test_throws ArgumentError add_arc!(b, vec_correct, vec_wrong)
+    @test_throws ArgumentError add_arc!(b, vec_wrong, vec_correct, vec_correct)
+    @test_throws ArgumentError add_arc!(b, vec_correct, vec_wrong, vec_correct)
+    @test_throws ArgumentError add_arc!(b, vec_correct, vec_correct, vec_wrong)
 
     b = Bloch()
     add_points!(b, [0.0, 0.0, 1.0])
@@ -147,12 +160,22 @@ end
         @test false
         @info "Render threw unexpected error" exception=e
     end
+
+    # test `state to Bloch vector` conversion and `add_states!` function
     b = Bloch()
-    x = basis(2, 0) + basis(2, 1)
-    y = basis(2, 0) - im * basis(2, 1)
-    z = basis(2, 0)
-    add_states!(b, [x, y, z])
-    th = range(0, 2π; length = 20);
+    Pauli_Ops = [sigmax(), sigmay(), sigmaz()]
+    ψ = rand_ket(2)
+    ρ = rand_dm(2)
+    x = basis(2, 0) + basis(2, 1)             # unnormalized Ket
+    ρ1 = 0.3 * rand_dm(2) + 0.4 * rand_dm(2)  # unnormalized density operator
+    ρ2 = Qobj(rand(ComplexF64, 2, 2))         # unnormalized and non-Hermitian Operator
+    add_states!(b, [ψ, ρ])
+    @test_logs (:warn,) (:warn,) (:warn,) (:warn,) add_states!(b, [x, ρ1, ρ2])
+    @test all(expect(Pauli_Ops, ψ) .≈ (b.vectors[1]))
+    @test all(expect(Pauli_Ops, ρ) .≈ (b.vectors[2]))
+    @test length(b.vectors) == 5
+
+    th = range(0, 2π; length = 20)
     xp = cos.(th);
     yp = sin.(th);
     zp = zeros(20);
@@ -174,7 +197,10 @@ end
     b = Bloch()
     ψ₁ = normalize(basis(2, 0) + basis(2, 1))
     ψ₂ = normalize(basis(2, 0) - im * basis(2, 1))
+    ψ₃ = basis(2, 0)
     add_line!(b, ψ₁, ψ₂; fmt = "r--")
+    add_arc!(b, ψ₁, ψ₂)
+    add_arc!(b, ψ₂, ψ₃, ψ₁)
     try
         fig, ax = render(b)
         @test !isnothing(fig)